perf(conv2): Optimize box blur 📦
Detects box kernels when computing a 2D convolution and decomposes them directly. That removes the
need for `dimfilter` and makes the logic between `ssim` and `originalSsim` more similar.

It is also now able to optimize all box kernels instead of the ones that were manually decomposed.

The internal implementation goes from `N^2` to `2N`. It should still be able to be
optimized further but this will do for now.

Closes #35, #46