fix(weberssim): fixes negative variance issues and overflows with proper rounding of the input
Current input luminance input is not rounded prior to the ssim function.  This means that we can't
use an integer implementation and expect the results to work properly until the numbers are
normalized.  As such, I've implemented a Uint8Array wrapper that rounds all of the luminance values
passed in.